fix useWindowWidth hook name
