+++ publisher/src/components/MetricConfiguration/Configuration.tsx
@@ -17,7 +17,7 @@
 
 import blueCheck from "@justice-counts/common/assets/status-check-icon.png";
 import { MIN_DESKTOP_WIDTH } from "@justice-counts/common/components/GlobalStyles";
-import useWindowWidth from "@justice-counts/common/hooks/useWIndowWidth";
+import { useWindowWidth } from "@justice-counts/common/hooks";
 import {
   SupervisionSubsystems,
   SupervisionSystem,
